D6: first draft "new scheme" ML mem stealing
Using a similar scheme to D7 changes, allow
stealing more early mem for ML.  See 200D/consts.h for
some notes and a pretty ASCII diagram.

D6 is a little different, one of the required constants
is encoded direct in an instruction, not loaded from mem.
So, we must add mov.w decoding and encoding to adjust this.

Also, because D6 has a large gap between firmware_entry
and cstart, RELOCSIZE is considerably higher, since we
block copy the whole region.

Currently, the effects / safety of pushing up sys_mem
by this larger size is untested.